Reattach a doc block, and give each scenario its own scratch [#200]
Inserting LocalePolicy between the SemanticInputs documentation and the struct left
that block documenting the enum instead - Doxygen associates only the immediately
preceding block with a declaration, which is the same failure mode CONTRIBUTING
records for a missing @file. The block sits above its struct again.

The temporary-directory helper mapped a fixed name straight under the system
temporary directory and removed it at construction as well as at destruction, so
two concurrent medui_tools_spec runs shared a path and either could delete files
the other was using. The failure that produces is an intermittent subprocess error
with nothing to do with the code under test.

It is fixed once rather than three times: the third copy of that class is what
prompted extracting it to tests/framework/TemporaryDirectory.hpp, and all three
carried the same defect. The name now takes a random suffix - std::random_device
rather than a process id, because that needs no platform header in a tree that
reaches the standard library through import std. Four now-redundant
create_directories calls went with it, since the shared class creates the directory
it owns.
